requirements.txt
Django==5.2
djangorestframework==3.16.0
Faker==37.1.0
python-decouple==3.8
